+++ src/js/tech/flash.js
@@ -21,21 +21,13 @@ let navigator = window.navigator;
  */
 class Flash extends Tech {
 
-  constructor(player, options, ready){
-    super(player, options, ready);
+  constructor(options, ready){
+    super(options, ready);
 
     let { source, parentEl } = options;
 
-    // Create a temporary element to be replaced by swf object
-    let placeHolder = this.el_ = Lib.createEl('div', { id: player.id() + '_temp_flash' });
-
     // Generate ID for swf object
-    let objId = player.id()+'_flash_api';
-
-    // Store player options in local var for optimization
-    // TODO: switch to using player methods instead of options
-    // e.g. player.autoplay();
-    let playerOptions = player.options_;
+    let objId = options.playerId+'_flash_api';
 
     // Merge default flashvars with ones passed in to init
     let flashVars = Lib.obj.merge({
@@ -46,25 +38,25 @@ class Flash extends Tech {
       'errorEventProxyFunction': 'videojs.Flash.onError',
 
       // Player Settings
-      'autoplay': playerOptions.autoplay,
-      'preload': playerOptions.preload,
-      'loop': playerOptions.loop,
-      'muted': playerOptions.muted
+      'autoplay': options.autoplay,
+      'preload': options.preload,
+      'loop': options.loop,
+      'muted': options.muted
 
-    }, options['flashVars']);
+    }, options.flashVars);
 
     // Merge default parames with ones passed in
     let params = Lib.obj.merge({
       'wmode': 'opaque', // Opaque is needed to overlay controls, but can affect playback performance
       'bgcolor': '#000000' // Using bgcolor prevents a white flash when the object is loading
-    }, options['params']);
+    }, options.params);
 
     // Merge default attributes with ones passed in
     let attributes = Lib.obj.merge({
       'id': objId,
       'name': objId, // Both ID and Name needed or swf to identify itself
       'class': 'vjs-tech'
-    }, options['attributes']);
+    }, options.attributes);
 
     // If source was supplied pass as a flash var.
     if (source) {
@@ -73,33 +65,15 @@ class Flash extends Tech {
       });
     }
 
-    // Add placeholder to player div
-    Lib.insertFirst(placeHolder, parentEl);
-
     // Having issues with Flash reloading on certain page actions (hide/resize/fullscreen) in certain browsers
     // This allows resetting the playhead when we catch the reload
-    if (options['startTime']) {
+    if (options.startTime) {
       this.ready(function(){
         this.load();
         this.play();
-        this['currentTime'](options['startTime']);
+        this.currentTime(options.startTime);
       });
     }
-
-    // firefox doesn't bubble mousemove events to parent. videojs/video-js-swf#37
-    // bugzilla bug: https://bugzilla.mozilla.org/show_bug.cgi?id=836786
-    if (Lib.IS_FIREFOX) {
-      this.ready(function(){
-        this.on('mousemove', function(){
-          // since it's a custom event, don't bubble higher than the player
-          this.player().trigger({ 'type':'mousemove', 'bubbles': false });
-        });
-      });
-    }
-
-    // native click events on the SWF aren't triggered on IE11, Win8.1RT
-    // use stageclick events triggered from inside the SWF instead
-    player.on('stageclick', player.reportUserActivity);
 
     window.videojs = window.videojs || {};
     window.videojs.Flash = window.videojs.Flash || {};
@@ -107,7 +81,8 @@ class Flash extends Tech {
     window.videojs.Flash.onEvent = Flash.onEvent;
     window.videojs.Flash.onError = Flash.onError;
 
-    this.el_ = Flash.embed(options['swf'], placeHolder, flashVars, params, attributes);
+    this.el_ = Flash.embed(options.swf, flashVars, params, attributes);
+    this.el_.tech = this;
   }
 
   play() {
@@ -120,7 +95,7 @@ class Flash extends Tech {
 
   src(src) {
     if (src === undefined) {
-      return this['currentSrc']();
+      return this.currentSrc();
     }
 
     // Setting src through `src` not `setSrc` will be deprecated
@@ -134,7 +109,7 @@ class Flash extends Tech {
 
     // Currently the SWF doesn't autoplay if you load a source later.
     // e.g. Load player w/ no source, wait 2s, set src.
-    if (this.player_.autoplay()) {
+    if (this.autoplay()) {
       var tech = this;
       this.setTimeout(function(){ tech.play(); }, 0);
     }
@@ -290,17 +265,13 @@ Flash.formats = {
 
 Flash.onReady = function(currSwf){
   let el = Lib.el(currSwf);
+  let tech = el && el.tech;
 
-  // get player from the player div property
-  const player = el && el.parentNode && el.parentNode['player'];
-
-  // if there is no el or player then the tech has been disposed
+  // if there is no el then the tech has been disposed
   // and the tech element was removed from the player div
-  if (player) {
-    // reference player on tech element
-    el['player'] = player;
+  if (tech && tech.el()) {
     // check that the flash object is really ready
-    Flash['checkReady'](player.tech);
+    Flash.checkReady(tech);
   }
 };
 
@@ -326,21 +297,21 @@ Flash.checkReady = function(tech){
 
 // Trigger events from the swf on the player
 Flash.onEvent = function(swfID, eventName){
-  let player = Lib.el(swfID)['player'];
-  player.trigger(eventName);
+  let tech = Lib.el(swfID).tech;
+  tech.trigger(eventName);
 };
 
 // Log errors from the swf
 Flash.onError = function(swfID, err){
-  const player = Lib.el(swfID)['player'];
+  const tech = Lib.el(swfID).tech;
   const msg = 'FLASH: '+err;
 
-  if (err == 'srcnotfound') {
-    player.error({ code: 4, message: msg });
+  if (err === 'srcnotfound') {
+    tech.trigger('error', { code: 4, message: msg });
 
   // errors we haven't categorized into the media errors
   } else {
-    player.error(msg);
+    tech.trigger('error', msg);
   }
 };
 
@@ -364,15 +335,12 @@ Flash.version = function(){
 };
 
 // Flash embedding method. Only used in non-iframe mode
-Flash.embed = function(swf, placeHolder, flashVars, params, attributes){
+Flash.embed = function(swf, flashVars, params, attributes){
   const code = Flash.getEmbedCode(swf, flashVars, params, attributes);
 
   // Get element by embedding code and retrieving created element
   const obj = Lib.createEl('div', { innerHTML: code }).childNodes[0];
 
-  const par = placeHolder.parentNode;
-
-  placeHolder.parentNode.replaceChild(obj, placeHolder);
   return obj;
 };